# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Marketplace Local API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stores/{storeId}/products:
    get:
      operationId: listStoreProducts
      tags: [products]
      summary: List products of a store
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product in a store
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products (for client browsing)
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: storeId
          in: query
          required: false
          schema:
            type: integer
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/{id}:
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product deleted
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        stock:
          type: integer
        storeId:
          type: integer
        storeName:
          type: string
        categoryId:
          type: integer
        imageUrl:
          type: string
        isAvailable:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - description
        - price
        - stock
        - storeId
        - isAvailable
        - createdAt
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        stock:
          type: integer
        imageUrl:
          type: string
        isAvailable:
          type: boolean
      required:
        - name
        - description
        - price
        - stock
...